pyproject.toml
[tool.poetry]
name = "event-driven-msa-with-flask"
version = "0.1.0"
description = ""
authors = ["do-huni <orangnlp@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
flask = "^3.0.3"
psycopg2 = "^2.9.9"
asyncpg = "^0.29.0"
databases = "^0.9.0"
aiokafka = "^0.11.0"
sqlalchemy = "^2.0.35"
pylance = "^0.18.2"
dependency-injector = "^4.42.0"
jsonschema = "^4.23.0"
python-dotenv = "^1.0.1"
flask-restx = "^1.3.0"
kafka-python = "^2.0.2"


[tool.poetry.group.dev.dependencies]
mypy = "^1.11.2"
pytest = "^8.3.3"
pytest-asyncio = "^0.24.0"
pytest-cov = "^5.0.0"
black = "^24.8.0"
isort = "^5.13.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
